requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Werkzeug==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
SQLAlchemy==1.4.23
requests==2.31.0
oauthlib==3.2.2
pyOpenSSL==23.2.0
psycopg2-binary==2.9.5
boto3==1.26.0
Google-Cloud-Storage==2.7.0